refactor: KDE: Make plugin_script_handler.rs
