+++ src/client/kde/mod.rs
@@ -1,4 +1,5 @@
 mod adhoc_script_handler;
 mod kde_client;
+mod plugin_script_handler;
 
 pub use kde_client::KdeClient;
